backend/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.0
idna==3.11
joblib==1.5.3
numpy==2.4.3
polyline==2.0.4
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
